Change TagApi for new structure
