+++ src/Api/TagApi.php
@@ -4,49 +4,46 @@ declare(strict_types=1);
 
 namespace DigitalCz\DigiSign\Api;
 
-use DigitalCz\DigiSign\Request\Recipient\RecipientDeleteRequest;
-use DigitalCz\DigiSign\Request\Tag\TagDeleteRequest;
-use DigitalCz\DigiSign\Request\Tag\TagGetRequest;
-use DigitalCz\DigiSign\Request\Tag\TagPostRequest;
-use DigitalCz\DigiSign\Request\Tag\TagPutRequest;
-use DigitalCz\DigiSign\Response\Recipient\RecipientDeleteResponse;
-use DigitalCz\DigiSign\Response\Tag\TagDeleteResponse;
-use DigitalCz\DigiSign\Response\Tag\TagResponse;
-use DigitalCz\DigiSign\ValueObject\Request\Tag as RequestTag;
-use DigitalCz\DigiSign\ValueObject\Response\Tag as ResponseTag;
+use DigitalCz\DigiSign\Http\Request\Tag\TagDeleteRequest;
+use DigitalCz\DigiSign\Http\Request\Tag\TagGetRequest;
+use DigitalCz\DigiSign\Http\Request\Tag\TagPostRequest;
+use DigitalCz\DigiSign\Http\Request\Tag\TagPutRequest;
+use DigitalCz\DigiSign\Http\Response\Tag\TagDeleteResponse;
+use DigitalCz\DigiSign\Http\Response\Tag\TagResponse;
+use DigitalCz\DigiSign\Model\DTO\TagData;
+use DigitalCz\DigiSign\Model\ValueObject\Tag;
 
 class TagApi extends BaseApi
 {
-    public function createTag(RequestTag $object): ResponseTag
+    public function createTag(TagData $object): Tag
     {
-        $httpRequest = (new TagPostRequest($this->httpRequestFactory, $this->httpStreamFactory))($object);
-        $httpResponse = $this->client->request($httpRequest);
+        $httpRequest = (new TagPostRequest($this->requestBuilder))($object);
+        $httpResponse = $this->client->sendRequest($httpRequest);
 
-        return (new TagResponse($httpResponse))();
+        return (new TagResponse())($httpResponse);
     }
 
-    public function updateTag(string $tagId, RequestTag $object): ResponseTag
+    public function updateTag(string $tagId, TagData $object): Tag
     {
-        $httpRequest = (new TagPutRequest($this->httpRequestFactory, $this->httpStreamFactory))($tagId, $object);
-        $httpResponse = $this->client->request($httpRequest);
+        $httpRequest = (new TagPutRequest($this->requestBuilder))($tagId, $object);
+        $httpResponse = $this->client->sendRequest($httpRequest);
 
-        return (new TagResponse($httpResponse))();
+        return (new TagResponse())($httpResponse);
     }
 
-    public function getTag(string $tagId): ResponseTag
+    public function getTag(string $tagId): Tag
     {
-        $httpRequest = (new TagGetRequest($this->httpRequestFactory, $this->httpStreamFactory))($tagId);
-        $httpResponse = $this->client->request($httpRequest);
+        $httpRequest = (new TagGetRequest($this->requestBuilder))($tagId);
+        $httpResponse = $this->client->sendRequest($httpRequest);
 
-        return (new TagResponse($httpResponse))();
+        return (new TagResponse())($httpResponse);
     }
 
     public function deleteTag(string $tagId): int
     {
-        $httpRequest =
-            (new TagDeleteRequest($this->httpRequestFactory, $this->httpStreamFactory))($tagId);
-        $httpResponse = $this->client->request($httpRequest);
+        $httpRequest = (new TagDeleteRequest($this->requestBuilder))($tagId);
+        $httpResponse = $this->client->sendRequest($httpRequest);
 
-        return (new TagDeleteResponse($httpResponse))();
+        return (new TagDeleteResponse())($httpResponse);
     }
 }
