+++ src/Http/Response/Tag/TagDeleteResponse.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Http\Response\Tag;
+
+use DigitalCz\DigiSign\Http\Response\BaseHttpCodeResponse;
+
+class TagDeleteResponse extends BaseHttpCodeResponse
+{
+
+}
