+++ src/Http/Response/Tag/TagResponse.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigitalCz\DigiSign\Http\Response\Tag;
+
+use DigitalCz\DigiSign\Http\Response\BaseHttpResponse;
+use DigitalCz\DigiSign\Model\ValueObject\Tag;
+use Psr\Http\Message\ResponseInterface;
+
+class TagResponse extends BaseHttpResponse
+{
+
+    public function __invoke(ResponseInterface $response): Tag
+    {
+        return Tag::fromArray($this->parseResponseBody($response));
+    }
+}
