+++ src/Model/ValueObject/Tag.php
@@ -2,24 +2,21 @@
 
 declare(strict_types=1);
 
-namespace DigitalCz\DigiSign\ValueObject\Request;
+namespace DigitalCz\DigiSign\Model\ValueObject;
+
+use DateTimeImmutable;
 
 class Tag
 {
     /**
      * @var string
      */
+    private $id;
+
+    /**
+     * @var string
+     */
     private $type;
-
-    /**
-     * @var string
-     */
-    private $recipient;
-
-    /**
-     * @var string
-     */
-    private $document;
 
     /**
      * @var int|null
@@ -36,20 +33,32 @@ class Tag
      */
     private $yPosition;
 
+    /**
+     * @var DateTimeImmutable
+     */
+    private $createdAt;
+
+    /**
+     * @var DateTimeImmutable
+     */
+    private $updatedAt;
+
     public function __construct(
+        string $id,
         string $type,
-        string $recipient,
-        string $document,
         ?int $page,
         ?int $xPosition,
-        ?int $yPosition
+        ?int $yPosition,
+        DateTimeImmutable $createdAt,
+        DateTimeImmutable $updatedAt
     ) {
+        $this->id = $id;
         $this->type = $type;
-        $this->recipient = $recipient;
-        $this->document = $document;
         $this->page = $page;
         $this->xPosition = $xPosition;
         $this->yPosition = $yPosition;
+        $this->createdAt = $createdAt;
+        $this->updatedAt = $updatedAt;
     }
 
     /**
@@ -58,12 +67,13 @@ class Tag
     public static function fromArray(array $data): self
     {
         return new self(
+            $data['id'],
             $data['type'],
-            $data['recipient'],
-            $data['document'],
             $data['page'] ? (int)$data['page'] : null,
             $data['xPosition'] ? (int)$data['xPosition'] : null,
-            $data['yPosition'] ? (int)$data['yPosition'] : null
+            $data['yPosition'] ? (int)$data['yPosition'] : null,
+            $data['createdAt'] ? new DateTimeImmutable($data['createdAt']) : null,
+            $data['updatedAt'] ? new DateTimeImmutable($data['updatedAt']) : null
         );
     }
 
@@ -73,28 +83,24 @@ class Tag
     public function toArray(): array
     {
         return [
+            'id' => $this->id,
             'type' => $this->type,
             'page' => $this->page,
             'xPosition' => $this->xPosition,
             'yPosition' => $this->yPosition,
-            'recipient' => $this->recipient,
-            'document' => $this->document,
+            'createdAt' => $this->createdAt ? $this->createdAt->format('c') : null,
+            'updatedAt' => $this->updatedAt ? $this->updatedAt->format('c') : null,
         ];
+    }
+
+    public function getId(): string
+    {
+        return $this->id;
     }
 
     public function getType(): string
     {
         return $this->type;
-    }
-
-    public function getRecipient(): string
-    {
-        return $this->recipient;
-    }
-
-    public function getDocument(): string
-    {
-        return $this->document;
     }
 
     public function getPage(): ?int
@@ -110,5 +116,15 @@ class Tag
     public function getYPosition(): ?int
     {
         return $this->yPosition;
+    }
+
+    public function getCreatedAt(): DateTimeImmutable
+    {
+        return $this->createdAt;
+    }
+
+    public function getUpdatedAt(): DateTimeImmutable
+    {
+        return $this->updatedAt;
     }
 }
