--- src/Response/Tag/TagDeleteResponse.php
@@ -1,12 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace DigitalCz\DigiSign\Response\Tag;
-
-use DigitalCz\DigiSign\Response\BaseHttpCodeResponse;
-
-class TagDeleteResponse extends BaseHttpCodeResponse
-{
-
-}
